Floral fertilizer can produce shimmering mushrooms
It spawns flowers on blocks in the `minecraft:dirt` tag (except in the nether), and mushrooms on blocks in the `minecraft:mushroom_grow_block` tag. If the block matches both conditions, flowers are preferred over mushrooms at a ratio of 2:1.

This commit also removes the random render offset for the mushrooms to match their interaction shape.